requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
django-countries==8.2.0
django-debug-toolbar==6.1.0
django-heroku==0.3.1
djangorestframework==3.16.1
idna==3.11
narwhals==2.15.0
numpy==2.4.1
packaging==25.0
pandas==2.3.3
pillow==12.1.0
plotly==6.5.2
psycopg2==2.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.11.0
xlsxwriter==3.2.9
xlwt==1.3.0
